+++ AIToolbox/experiment_save/experiment_saver.py
@@ -2,10 +2,10 @@ from abc import ABC, abstractmethod
 import time
 import datetime
 
-from AIToolbox.AWS.model_save import KerasS3ModelSaver, PyTorchS3ModelSaver
-from AIToolbox.AWS.results_save import S3ResultsSaver
-from AIToolbox.GoogleCloud.model_save import KerasGoogleStorageModelSaver, PyTorchGoogleStorageModelSaver
-from AIToolbox.GoogleCloud.results_save import GoogleStorageResultsSaver
+from AIToolbox.cloud.AWS.model_save import KerasS3ModelSaver, PyTorchS3ModelSaver
+from AIToolbox.cloud.AWS.results_save import S3ResultsSaver
+from AIToolbox.cloud.GoogleCloud.model_save import KerasGoogleStorageModelSaver, PyTorchGoogleStorageModelSaver
+from AIToolbox.cloud.GoogleCloud.results_save import GoogleStorageResultsSaver
 
 
 class AbstractExperimentSaver(ABC):
@@ -35,7 +35,7 @@ class BaseFullExperimentS3Saver(AbstractExperimentSaver):
         """
 
         Args:
-            model_saver (AIToolbox.AWS.model_save.AbstractModelSaver):
+            model_saver (AIToolbox.cloud.AWS.model_save.AbstractModelSaver):
             project_name (str):
             experiment_name (str):
             bucket_name (str):
@@ -127,7 +127,7 @@ class BaseFullExperimentGoogleStorageSaver(BaseFullExperimentS3Saver):
         """
 
         Args:
-            model_saver (AIToolbox.AWS.model_save.AbstractModelSaver):
+            model_saver (AIToolbox.cloud.AWS.model_save.AbstractModelSaver):
             project_name (str):
             experiment_name (str):
             bucket_name (str):
