+++ AIToolbox/torchtrain/callbacks/callbacks.py
@@ -1,5 +1,5 @@
-from AIToolbox.AWS.model_save import PyTorchS3ModelSaver
-from AIToolbox.GoogleCloud.model_save import PyTorchGoogleStorageModelSaver
+from AIToolbox.cloud.AWS.model_save import PyTorchS3ModelSaver
+from AIToolbox.cloud.GoogleCloud.model_save import PyTorchGoogleStorageModelSaver
 from AIToolbox.experiment_save.local_model_save import PyTorchLocalModelSaver
 from AIToolbox.experiment_save.experiment_saver import FullPyTorchExperimentS3Saver, FullPyTorchExperimentGoogleStorageSaver
 from AIToolbox.experiment_save.local_experiment_saver import FullPyTorchExperimentLocalSaver
