+++ tests/test_torchtrain/test_callbacks.py
@@ -5,10 +5,9 @@ from tests.utils import *
 from AIToolbox.torchtrain.callbacks.callbacks import AbstractCallback, ModelCheckpointCallback, ModelTrainEndSaveCallback, \
     EarlyStoppingCallback
 from AIToolbox.torchtrain.train_loop import TrainLoop
-from AIToolbox.AWS.model_save import PyTorchS3ModelSaver
-from AIToolbox.GoogleCloud.model_save import PyTorchGoogleStorageModelSaver
+from AIToolbox.cloud.AWS.model_save import PyTorchS3ModelSaver
 from AIToolbox.experiment_save.local_model_save import PyTorchLocalModelSaver
-from AIToolbox.experiment_save.experiment_saver import FullPyTorchExperimentS3Saver, FullPyTorchExperimentGoogleStorageSaver
+from AIToolbox.experiment_save.experiment_saver import FullPyTorchExperimentS3Saver
 from AIToolbox.experiment_save.local_experiment_saver import FullPyTorchExperimentLocalSaver
